Create mockup about.html
This is in essence what the new "About" page should look like. However, it is unclear how exactly the DSpace JSPUI constructs the "About" page, so this file may need to be revised--perhaps to include just the body. It is also unclear whether JSPUI will reference "styles.css," and if so, where that is in relation to this file's directory.