Create mockup about-body
I suspect that the DSpace JSPUI takes the content of the "About" page as an option value inserted into the "Messages.properties" file. In that case, we don't want a full HTML page, as in the "about.html" file in the mockup folder here. Instead, we would just want the static content wrapped in <div></div> tags, so it can be easily dropped into the HTML page JSPUI automatically constructs for the viewer. I'm not certain about the stylesheets used in that case, but I believe what I have here will look acceptable regardless.